Started (just barely) to turn this into a module
I don't have a lot of experience doing this, so please be patient.

For this to work, the path to the newly created vtoolbox directory
needs to be in your PYTHONPATH. In know how to do this in unix, not
windows (I'm guessing you use windows?). When this is an installed
module, this won't be an issue as pip makes this all work. For
development we have to manually add it.

If you have a better construct that will work, please advise and
improve.